Use a GADT to reduce boilerplate to print results.
Day 10 had two parts with different result types, an int and printed text. The
harness required both parts of a given day to return the same concrete type, so
I had to change part a to print its result as text.

I've changed the harness to require days to return a GADT which can print ints,
strings, or units (for things that print on their own), substantially reducing
boilerplate and the costs of having two parts with different return types. It
does however require me to include [Printer.of_int] and friends at the end of
every part.